[차소연] 재료 직접 등록, 전체 조회, 목록 수정, 삭제 기능 구현

+++ src/main/java/CookSave/CookSaveback/Icon/domain/Icon.java
@@ -12,7 +12,7 @@ public class Icon {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(nullable = false, updatable = false)
-    private Long iconId;
+    private Integer iconId;
 
     @Column(nullable = false)
     private String image;  // 아이콘 이미지 URL
